AS-442-fixes Organization data validation fixes + element delete function
- [x] Total expenditure validation fix + delete
- [x] Recipient country budget fix + delete
- [x] Recipient region budget fix + delete
- [x] Recipient organization budget fix + delete
- [x] Total budget validation fix + delete
- [x] Organization name delete

+++ app/Http/Controllers/Complete/Activity/ActivityController.php
@@ -743,7 +743,13 @@ class ActivityController extends Controller
     public function deleteElement($id, $element)
     {
         $activity = $this->activityManager->getActivityData($id);
-        $result   = $this->activityManager->deleteElement($activity, $element);
+
+        if (Gate::denies('ownership', $activity)) {
+            return redirect()->back()->withResponse($this->getNoPrivilegesMessage());
+        }
+
+        $result = $this->activityManager->deleteElement($activity, $element);
+
         if ($result) {
             $this->activityManager->resetActivityWorkflow($id);
             $response = ['type' => 'success', 'code' => ['activity_element_removed', ['element' => 'activity']]];
